Se completaron las implementaciones de los metodos CRUD de pago. Ademas se agregaron validaciones de reglas de negocio

+++ src/main/java/mx/uam/tsis2020/prestamax/negocio/modelo/Pago.java
@@ -50,5 +50,5 @@ public class Pago {
 	
 	@NotNull
 	@ApiModelProperty(notes="penalizacion del pago tardio", required=true)
-	private Integer penalizacion;
+	private Double penalizacion;
 }
